extract function in load data
